Working on a default logger
