+++ dawn/test/unit-test/dawn/CodeGen/Cuda/TestMain.cpp
@@ -13,16 +13,11 @@
 //===------------------------------------------------------------------------------------------===//
 
 #include "dawn/Support/STLExtras.h"
-#include "dawn/Unittest/UnittestLogger.h"
 #include <gtest/gtest.h>
 
 int main(int argc, char* argv[]) {
   // Initialize gtest
   testing::InitGoogleTest(&argc, argv);
-
-  // Initialize Unittest-Logger
-  auto logger = std::make_unique<dawn::UnittestLogger>();
-  dawn::Logger::getSingleton().registerLogger(logger.get());
 
   return RUN_ALL_TESTS();
 }
