+++ dawn/test/unit-test/dawn/CodeGen/Naive/TestCodeGenNaive.cpp
@@ -21,7 +21,6 @@
 #include "dawn/Support/FileUtil.h"
 #include "dawn/Unittest/CompilerUtil.h"
 #include "dawn/Unittest/IIRBuilder.h"
-#include "dawn/Unittest/UnittestLogger.h"
 
 #include <fstream>
 #include <gtest/gtest.h>
