+++ dawn/test/unit-test/dawn/IIR/TestMain.cpp
@@ -13,17 +13,12 @@
 //===------------------------------------------------------------------------------------------===//
 
 #include "dawn/Support/STLExtras.h"
-#include "dawn/Unittest/UnittestLogger.h"
 #include <gtest/gtest.h>
 
 int main(int argc, char* argv[]) {
 
   // Initialize gtest
   testing::InitGoogleTest(&argc, argv);
-
-  // Initialize Unittest-Logger
-  auto logger = std::make_unique<dawn::UnittestLogger>();
-  dawn::Logger::getSingleton().registerLogger(logger.get());
 
   return RUN_ALL_TESTS();
 }
